feat: move testing artifact writing into run_testing() for pipeline-mode support
